requirements.txt
streamlit
tensorflow-cpu   
numpy
opencv-python-headless
Pillow
gdown    
streamlit-drawable-canvas       